+++ src/Detach/Utils/MathUtils.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using System.Runtime.CompilerServices;
 
 namespace Detach.Utils;
@@ -9,17 +10,43 @@ public static class MathUtils
 
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static float ToDegrees(float radians)
-		=> radians * _toDeg;
+	{
+		return radians * _toDeg;
+	}
 
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static float ToRadians(float degrees)
-		=> degrees * _toRad;
+	{
+		return degrees * _toRad;
+	}
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	public static Vector3 ToRadians(Vector3 degrees)
+	{
+		return degrees * MathF.PI / 180;
+	}
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	public static Vector3 ToDegrees(Vector3 radians)
+	{
+		return radians * 180 / MathF.PI;
+	}
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	public static Matrix4x4 CreateRotationMatrixFromEulerAngles(Vector3 eulerAngles)
+	{
+		return Matrix4x4.CreateFromYawPitchRoll(eulerAngles.Y, eulerAngles.X, eulerAngles.Z);
+	}
 
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static float Lerp(float value1, float value2, float amount)
-		=> value1 + (value2 - value1) * amount;
+	{
+		return value1 + (value2 - value1) * amount;
+	}
 
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public static bool IsFloatReal(float value)
-		=> !float.IsNaN(value) && !float.IsInfinity(value);
+	{
+		return !float.IsNaN(value) && !float.IsInfinity(value);
+	}
 }