+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.0
+
+### Added
+
+- Added `MathUtils.ToRadians` and `MathUtils.ToDegrees` overloads accepting `Vector3`.
+- Added `MathUtils.CreateRotationMatrixFromEulerAngles` method.
+
 ## 0.1.0
 
 - Initial release